move mockInstance to avoid loading stuff in unit tests
