+++ packages/cli/test/integration/ExternalSecrets/externalSecrets.api.test.ts
@@ -1,36 +1,38 @@
 import type { SuperAgentTest } from 'supertest';
 import { License } from '@/License';
-import * as utils from '../shared/utils/';
 import type { ExternalSecretsSettings, SecretsProviderState } from '@/Interfaces';
 import { Cipher } from 'n8n-core';
 import { SettingsRepository } from '@db/repositories/settings.repository';
 import { Container } from 'typedi';
 import { ExternalSecretsProviders } from '@/ExternalSecrets/ExternalSecretsProviders.ee';
+import config from '@/config';
+import { ExternalSecretsManager } from '@/ExternalSecrets/ExternalSecretsManager.ee';
+import { CREDENTIAL_BLANKING_VALUE } from '@/constants';
+import { jsonParse, type IDataObject } from 'n8n-workflow';
+import { mock } from 'jest-mock-extended';
+
+import { mockInstance } from '../../shared/mocking';
+import { setupTestServer } from '../shared/utils';
+import { createOwner, createUser } from '../shared/db/users';
 import {
 	DummyProvider,
 	FailedProvider,
 	MockProviders,
 	TestFailProvider,
 } from '../../shared/ExternalSecrets/utils';
-import config from '@/config';
-import { ExternalSecretsManager } from '@/ExternalSecrets/ExternalSecretsManager.ee';
-import { CREDENTIAL_BLANKING_VALUE } from '@/constants';
-import { jsonParse, type IDataObject } from 'n8n-workflow';
-import { mock } from 'jest-mock-extended';
-import { createOwner, createUser } from '../shared/db/users';
 
 let authOwnerAgent: SuperAgentTest;
 let authMemberAgent: SuperAgentTest;
 
-const licenseLike = utils.mockInstance(License, {
+const licenseLike = mockInstance(License, {
 	isExternalSecretsEnabled: jest.fn().mockReturnValue(true),
 	isWithinUsersLimit: jest.fn().mockReturnValue(true),
 });
 
 const mockProvidersInstance = new MockProviders();
-utils.mockInstance(ExternalSecretsProviders, mockProvidersInstance);
+mockInstance(ExternalSecretsProviders, mockProvidersInstance);
 
-const testServer = utils.setupTestServer({ endpointGroups: ['externalSecrets'] });
+const testServer = setupTestServer({ endpointGroups: ['externalSecrets'] });
 
 const connectedDate = '2023-08-01T12:32:29.000Z';
