+++ packages/cli/test/integration/audit/nodes.risk.test.ts
@@ -2,14 +2,15 @@ import { v4 as uuid } from 'uuid';
 import { Container } from 'typedi';
 import { audit } from '@/audit';
 import { OFFICIAL_RISKY_NODE_TYPES, NODES_REPORT } from '@/audit/constants';
-import { getRiskSection, MOCK_PACKAGE, saveManualTriggerWorkflow } from './utils';
-import * as testDb from '../shared/testDb';
 import { toReportTitle } from '@/audit/utils';
-import { mockInstance } from '../shared/utils/';
 import { LoadNodesAndCredentials } from '@/LoadNodesAndCredentials';
 import { NodeTypes } from '@/NodeTypes';
 import { CommunityPackagesService } from '@/services/communityPackages.service';
 import { WorkflowRepository } from '@db/repositories/workflow.repository';
+
+import { mockInstance } from '../../shared/mocking';
+import { getRiskSection, MOCK_PACKAGE, saveManualTriggerWorkflow } from './utils';
+import * as testDb from '../shared/testDb';
 
 const nodesAndCredentials = mockInstance(LoadNodesAndCredentials);
 nodesAndCredentials.getCustomDirectories.mockReturnValue([]);
