+++ packages/cli/test/integration/binaryData.api.test.ts
@@ -1,8 +1,10 @@
 import fsp from 'node:fs/promises';
 import { Readable } from 'node:stream';
 import { BinaryDataService, FileNotFoundError } from 'n8n-core';
-import { mockInstance, setupTestServer } from './shared/utils';
 import type { SuperAgentTest } from 'supertest';
+
+import { mockInstance } from '../shared/mocking';
+import { setupTestServer } from './shared/utils';
 import { createOwner } from './shared/db/users';
 
 jest.mock('fs/promises');
