+++ packages/cli/test/integration/commands/worker.cmd.test.ts
@@ -1,4 +1,3 @@
-import { mockInstance } from '../shared/utils/';
 import { Worker } from '@/commands/worker';
 import * as Config from '@oclif/config';
 import config from '@/config';
@@ -17,6 +16,8 @@ import { PostHogClient } from '@/posthog';
 import { RedisService } from '@/services/redis.service';
 import { OrchestrationHandlerWorkerService } from '@/services/orchestration/worker/orchestration.handler.worker.service';
 import { OrchestrationWorkerService } from '@/services/orchestration/worker/orchestration.worker.service';
+
+import { mockInstance } from '../../shared/mocking';
 
 const oclifConfig: Config.IConfig = new Config.Config({ root: __dirname });
