+++ packages/cli/test/integration/community-packages.api.test.ts
@@ -7,14 +7,9 @@ import { LoadNodesAndCredentials } from '@/LoadNodesAndCredentials';
 import { Push } from '@/push';
 import { CommunityPackagesService } from '@/services/communityPackages.service';
 
+import { mockInstance } from '../shared/mocking';
 import { COMMUNITY_PACKAGE_VERSION } from './shared/constants';
-import {
-	mockInstance,
-	setupTestServer,
-	mockPackage,
-	mockNode,
-	mockPackageName,
-} from './shared/utils';
+import { setupTestServer, mockPackage, mockNode, mockPackageName } from './shared/utils';
 import { createOwner } from './shared/db/users';
 
 const communityPackagesService = mockInstance(CommunityPackagesService, {
