+++ packages/cli/test/integration/variables.test.ts
@@ -5,6 +5,8 @@ import { VariablesRepository } from '@db/repositories/variables.repository';
 import { generateNanoId } from '@db/utils/generators';
 import { License } from '@/License';
 import { VariablesService } from '@/environments/variables/variables.service';
+
+import { mockInstance } from '../shared/mocking';
 import * as testDb from './shared/testDb';
 import * as utils from './shared/utils/';
 import { createOwner, createUser } from './shared/db/users';
@@ -47,7 +49,7 @@ async function getVariableById(id: string) {
 }
 
 beforeAll(async () => {
-	utils.mockInstance(License, licenseLike);
+	mockInstance(License, licenseLike);
 
 	const owner = await createOwner();
 	authOwnerAgent = testServer.authAgentFor(owner);
