+++ packages/cli/test/shared/mocking.ts
