+++ packages/cli/test/unit/CredentialsHelper.test.ts
@@ -14,7 +14,7 @@ import { NodeTypes } from '@/NodeTypes';
 import { LoadNodesAndCredentials } from '@/LoadNodesAndCredentials';
 import { CredentialsRepository } from '@db/repositories/credentials.repository';
 import { SharedCredentialsRepository } from '@db/repositories/sharedCredentials.repository';
-import { mockInstance } from '../integration/shared/utils';
+import { mockInstance } from '../shared/mocking';
 
 describe('CredentialsHelper', () => {
 	mockInstance(CredentialsRepository);
