+++ packages/cli/test/unit/Telemetry.test.ts
@@ -4,8 +4,8 @@ import config from '@/config';
 import { flushPromises } from './Helpers';
 import { PostHogClient } from '@/posthog';
 import { mock } from 'jest-mock-extended';
-import { mockInstance } from '../integration/shared/utils';
 import { InstanceSettings } from 'n8n-core';
+import { mockInstance } from '../shared/mocking';
 
 jest.unmock('@/telemetry');
 jest.mock('@/license/License.service', () => {
