+++ packages/cli/test/unit/WorkflowExecuteAdditionalData.test.ts
@@ -1,7 +1,7 @@
 import { Container } from 'typedi';
 import { ExecutionMetadataRepository } from '@db/repositories/executionMetadata.repository';
 import { ExecutionMetadataService } from '@/services/executionMetadata.service';
-import { mockInstance } from '../integration/shared/utils';
+import { mockInstance } from '../shared/mocking';
 
 describe('WorkflowExecuteAdditionalData', () => {
 	const repository = mockInstance(ExecutionMetadataRepository);
