+++ packages/cli/test/unit/WorkflowRunner.test.ts
@@ -1,12 +1,13 @@
 import type { User } from '@db/entities/User';
-import * as testDb from '../integration/shared/testDb';
-import * as utils from '../integration/shared/utils/';
 import { WorkflowRunner } from '@/WorkflowRunner';
 import { WorkflowHooks, type ExecutionError, type IWorkflowExecuteHooks } from 'n8n-workflow';
 import { Push } from '@/push';
-import { mockInstance } from '../integration/shared/utils';
 import Container from 'typedi';
 import config from '@/config';
+
+import { mockInstance } from '../shared/mocking';
+import * as testDb from '../integration/shared/testDb';
+import { setupTestServer } from '../integration/shared/utils';
 import { getGlobalOwnerRole } from '../integration/shared/db/roles';
 import { createUser } from '../integration/shared/db/users';
 import { createWorkflow } from '../integration/shared/db/workflows';
@@ -15,7 +16,7 @@ import { createExecution } from '../integration/shared/db/executions';
 let owner: User;
 let runner: WorkflowRunner;
 let hookFunctions: IWorkflowExecuteHooks;
-utils.setupTestServer({ endpointGroups: [] });
+setupTestServer({ endpointGroups: [] });
 
 class Watchers {
 	workflowExecuteAfter = jest.fn();
