+++ packages/cli/test/unit/controllers/oAuth1Credential.controller.test.ts
@@ -16,7 +16,7 @@ import { VariablesService } from '@/environments/variables/variables.service';
 import { SecretsHelper } from '@/SecretsHelpers';
 import { CredentialsHelper } from '@/CredentialsHelper';
 
-import { mockInstance } from '../../integration/shared/utils';
+import { mockInstance } from '../../shared/mocking';
 
 describe('OAuth1CredentialController', () => {
 	mockInstance(Logger);
