+++ packages/cli/test/unit/push/websocket.push.test.ts
@@ -5,7 +5,7 @@ import { WebSocketPush } from '@/push/websocket.push';
 import { Logger } from '@/Logger';
 import type { User } from '@db/entities/User';
 import type { PushDataExecutionRecovered } from '@/Interfaces';
-import { mockInstance } from '../../integration/shared/utils';
+import { mockInstance } from '../../shared/mocking';
 
 jest.useFakeTimers();
