+++ packages/cli/test/unit/services/redis.service.test.ts
@@ -2,7 +2,7 @@ import Container from 'typedi';
 import { Logger } from '@/Logger';
 import config from '@/config';
 import { RedisService } from '@/services/redis.service';
-import { mockInstance } from '../../integration/shared/utils';
+import { mockInstance } from '../../shared/mocking';
 
 mockInstance(Logger);
 const redisService = Container.get(RedisService);
