+++ packages/cli/test/unit/services/user.service.test.ts
@@ -5,7 +5,7 @@ import config from '@/config';
 import { User } from '@db/entities/User';
 import { UserRepository } from '@db/repositories/user.repository';
 import { UserService } from '@/services/user.service';
-import { mockInstance } from '../../integration/shared/utils';
+import { mockInstance } from '../../shared/mocking';
 
 describe('UserService', () => {
 	config.set('userManagement.jwtSecret', 'random-secret');
